Fixes Requirements for Vim Plugins
This adds some missing reqs that was causing vim plugins to not be
installed properly. Since they are installed to work with Pathogen, the
Pathogen project needs to be a requisite of the .vim/bundle directory
into which they are installed. This fixes that issue and makes
.vim/bundle a requirement for each git project installed into that
folder. Though it might be better in the future to use require_in under
the state definition for bundle.